Update index.html
Here is a suggestion 

The text "about" is unclear without context and may be confusing to screen readers. Consider rearranging the <a></a> tags or including special screen reader text.
Relevant code:
<a href="#about" style="color: rgb(226, 87, 76); background-color: rgba(0, 0, 0, 0);">about</a>

Perhaps make the link into a button?

For the input, screen-reader my not be able to read the place holder. the suggestion is to create a label

Also check colour contrast for your about link